Changes to router with no data
